refactor: extract input_endpoint.rs
